requirements.txt
flask
requests
python-docx
PyPDF2
gunicorn
